setup group colors

+++ composeApp/src/commonMain/kotlin/com/trm/alarmist/feature/group/GroupState.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class GroupState(
   val name: String? = "",
-  val color: Int? = null,
+  val color: Long? = null,
   val ungroupedAlarms: List<AlarmListModel> = emptyList(),
   val selectedAlarmIds: Set<Long> = emptySet(),
 )
